feat: Remove general lexeme field in `Token`
Move all needed lexemes from the `Token` type to the `TokenType` enum.
This helps to save memory from unneccesarily storing derivable lexemes for
tokens like symbols and keywords. This is essentially free, since the `TokenType`
enum is already storing a `String` for the `TokenType::String` variant,
so this should not increase the enum size.